+++ services/autoscaling/src/simcore_service_autoscaling/utils/utils_ec2.py
